测试将返回的data里面的title的value翻译
