+++ src/main/java/com/bogdatech/integration/ShopifyHttpIntegration.java
@@ -51,7 +51,7 @@ public class ShopifyHttpIntegration {
     public JSONObject getInfoByShopify(ShopifyRequest shopifyRequest, String query) {
         String string = sendShopifyPost(shopifyRequest, query);
         JSONObject jsonObject = JSONObject.parseObject(string);
-        return jsonObject;
+        return jsonObject.getJSONObject("data");
     }
 }
