+++ src/main/java/com/bogdatech/logic/TranslateService.java
@@ -1,19 +1,18 @@
 package com.bogdatech.logic;
 
 
+import com.alibaba.fastjson.JSONObject;
 import com.bogdatech.entity.TranslatesDO;
-import com.bogdatech.repository.JdbcRepository;
 import com.bogdatech.integration.TranslateApiIntegration;
 import com.bogdatech.model.controller.request.TranslateRequest;
 import com.bogdatech.model.controller.response.BaseResponse;
+import com.bogdatech.repository.JdbcRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
-
-import static com.bogdatech.enums.ErrorEnum.*;
+import static com.bogdatech.enums.ErrorEnum.TRANSLATE_ERROR;
 
 @Component
 @EnableAsync
@@ -47,29 +46,7 @@ public class TranslateService {
         return new BaseResponse<>().CreateErrorResponse(TRANSLATE_ERROR);
     }
 
-    public BaseResponse insertShopTranslateInfo(TranslateRequest request) {
-        String sql = "INSERT INTO Translates (shop_name, access_token, source, target) VALUES (?, ?, ?, ?)";
-        Object[] info = {request.getShopName(), request.getAccessToken(), request.getSource(), request.getTarget()};
-        int result = jdbcRepository.CUDInfo(info, sql);
-        if (result > 0) {
-            return new BaseResponse().CreateSuccessResponse(result);
-        }
-        return new BaseResponse<>().CreateErrorResponse(SQL_INSERT_ERROR);
-    }
 
-    public List<TranslatesDO> readTranslateInfo(int status){
-        String sql = "SELECT id,source,target,shop_name,status,create_at,update_at FROM Translates WHERE status = ?";
-        Object[] info = {status};
-        List<TranslatesDO> list =  jdbcRepository.readInfo(info, sql, TranslatesDO.class);
-        return list;
-    }
-
-    public int updateTranslateStatus(int id, int status){
-        String sql = "UPDATE Translates SET status = ? WHERE id = ?";
-        Object[] info = {status, id};
-        int result = jdbcRepository.CUDInfo(info, sql);
-        return result;
-    }
 
     @Async
     public void test(TranslatesDO request){
@@ -82,17 +59,24 @@ public class TranslateService {
         }
         System.out.println("翻译完成" + Thread.currentThread().getName());
         //更新状态
-        updateTranslateStatus(request.getId(), 0);
+        jdbcRepository.updateTranslateStatus(request.getId(), 0);
     }
 
-    public BaseResponse readInfoByShopName(TranslateRequest request) {
-        String sql = "SELECT id,source,target,shop_name,status,create_at,update_at FROM Translates WHERE shop_name = ?";
-        Object[] info = {request.getShopName()};
-        List<TranslatesDO> translatesDOS = jdbcRepository.readInfo(info, sql, TranslatesDO.class);
-        if (translatesDOS.size() > 0){
-            return new BaseResponse().CreateSuccessResponse(translatesDOS);
+    public BaseResponse userBDTranslateProduct(JSONObject object){
+        //用循环将object中名字为title的翻译
+        for (String key : object.keySet()) {
+            if ("title".equals(key)) {
+                String title = object.getString(key);
+                String result = translateApiIntegration.baiDuTranslate(new TranslateRequest(0,null,null, "en", "zh",title));
+                if (result!= null) {
+                    object.put(key, result);
+                } else {
+                    return new BaseResponse<>().CreateErrorResponse(TRANSLATE_ERROR);
+                }
+            }
         }
-           return new BaseResponse<>().CreateErrorResponse(SQL_SELECT_ERROR);
+        return new BaseResponse<>().CreateSuccessResponse(object);
     }
+
 }
