+++ src/main/java/com/bogdatech/repository/JdbcRepository.java
@@ -1,5 +1,8 @@
 package com.bogdatech.repository;
 
+import com.bogdatech.entity.TranslatesDO;
+import com.bogdatech.model.controller.request.TranslateRequest;
+import com.bogdatech.model.controller.response.BaseResponse;
 import com.bogdatech.utils.CamelToUnderscore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -10,6 +13,9 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
+
+import static com.bogdatech.enums.ErrorEnum.SQL_INSERT_ERROR;
+import static com.bogdatech.enums.ErrorEnum.SQL_SELECT_ERROR;
 
 @Component
 public class JdbcRepository {
@@ -56,5 +62,39 @@ public class JdbcRepository {
             throw new RuntimeException(e);
         }
 
+    }
+
+    public BaseResponse insertShopTranslateInfo(TranslateRequest request) {
+        String sql = "INSERT INTO Translates (shop_name, access_token, source, target) VALUES (?, ?, ?, ?)";
+        Object[] info = {request.getShopName(), request.getAccessToken(), request.getSource(), request.getTarget()};
+        int result = CUDInfo(info, sql);
+        if (result > 0) {
+            return new BaseResponse().CreateSuccessResponse(result);
+        }
+        return new BaseResponse<>().CreateErrorResponse(SQL_INSERT_ERROR);
+    }
+
+    public List<TranslatesDO> readTranslateInfo(int status){
+        String sql = "SELECT id,source,target,shop_name,status,create_at,update_at FROM Translates WHERE status = ?";
+        Object[] info = {status};
+        List<TranslatesDO> list =  readInfo(info, sql, TranslatesDO.class);
+        return list;
+    }
+
+    public int updateTranslateStatus(int id, int status){
+        String sql = "UPDATE Translates SET status = ? WHERE id = ?";
+        Object[] info = {status, id};
+        int result = CUDInfo(info, sql);
+        return result;
+    }
+
+    public BaseResponse readInfoByShopName(TranslateRequest request) {
+        String sql = "SELECT id,source,target,shop_name,status,create_at,update_at FROM Translates WHERE shop_name = ?";
+        Object[] info = {request.getShopName()};
+        List<TranslatesDO> translatesDOS = readInfo(info, sql, TranslatesDO.class);
+        if (translatesDOS.size() > 0){
+            return new BaseResponse().CreateSuccessResponse(translatesDOS);
+        }
+        return new BaseResponse<>().CreateErrorResponse(SQL_SELECT_ERROR);
     }
 }
