+++ src/main/java/com/bogdatech/task/TranslateTask.java
@@ -2,6 +2,7 @@ package com.bogdatech.task;
 
 import com.bogdatech.entity.TranslatesDO;
 import com.bogdatech.logic.TranslateService;
+import com.bogdatech.repository.JdbcRepository;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.EnableAsync;
@@ -17,6 +18,9 @@ import java.util.List;
 public class TranslateTask {
 
     @Autowired
+    private JdbcRepository jdbcRepository;
+
+    @Autowired
     private TranslateService translateService;
 
     @PostConstruct
@@ -24,7 +28,7 @@ public class TranslateTask {
     public void translate() {
         System.out.println("translate");
         //查询Translates表status为1的数据
-        List<TranslatesDO> list = translateService.readTranslateInfo(1);
+        List<TranslatesDO> list = jdbcRepository.readTranslateInfo(1);
         //遍历list，调用translateService.translate()方法
         for (TranslatesDO request : list) {
             System.out.println("request: " + request);
@@ -32,7 +36,7 @@ public class TranslateTask {
             System.out.println("正在进行翻译");
             translateService.test(request);
             //修改status状态
-            int i = translateService.updateTranslateStatus(request.getId(), 2);
+            int i = jdbcRepository.updateTranslateStatus(request.getId(), 2);
             //判断是否修改成功
             if (i > 0) {
                 System.out.println("修改成功");
